+++ .dockerignore
@@ -1,0 +1,7 @@
+__pycache__
+*.pyc
+*.pyo
+*.pyd
+*.log
+*.md
+Makefile
