+++ api/auth.py
@@ -4,7 +4,7 @@ from functools import wraps
 
 from flask import request
 
-from error_handlers import _make_response
+from api.error_handlers import _make_response
 
 
 FORTNITE_REPLAY_ELIM_API_AUTH_DIGEST = os.getenv("FORTNITE_REPLAY_ELIM_API_AUTH_DIGEST")
