+++ core/images/prompt_example.png
