+++ core/images/provide_username_example.png
