+++ core/images/ranks/diamond_three.png
