+++ core/images/ranks/diamond_two.png
