+++ core/images/ranks/diamong_one.png
