+++ core/images/ranks/gold_two.png
