+++ core/images/stats_diff_example.png
