+++ core/logger.py
@@ -103,4 +103,5 @@ def _log_response(response):
 
 def _is_healthcheck():
     """ Returns True if the request is a healthcheck request """
-    return request.path == "/fortnite/healthcheck"
+    # return request.path == "/fortnite/healthcheck"
+    return False
