+++ core/utils/globals.py
@@ -1,0 +1,3 @@
+# Temporary global variables until Replays Watcher v2 is released
+eliminated_by_me_dict = None
+eliminated_me_dict = None
